refactor: rename permits.js
